addressed comments after the pull request review
