+++ lesson9-oop-solid/src/equipment.ts
@@ -1,13 +1,12 @@
-import { IEquipment } from "./interfaces";
+// Abstract class representing manufacturing equipment.
+export abstract class Equipment {
+    protected name: string; // Protected: Only accessible within this class and subclasses.
 
-// Abstract class that provides a common structure for all equipment.
-export abstract class Equipment implements IEquipment {
-  name: string;
+    public constructor(name: string) {
+        this.name = name;
+    }
 
-  constructor(name: string) {
-    this.name = name;
-  }
-
-  // Abstract method that each subclass must implement.
-  abstract operate(): Promise<void>;
+    // Abstract method that must be implemented by subclasses.
+    public abstract operate(): Promise<void>;
 }
+
